+++ src/tr2/game/console/cmd/easy_config.c
@@ -13,6 +13,7 @@ static COMMAND_TO_OPTION_MAP m_CommandToOptionMap[] = {
     { "cheats", &g_Config.gameplay.enable_cheats },
     { "wireframe", &g_Config.rendering.enable_wireframe },
     { "fps", &g_Config.rendering.fps },
+    { "lighting", &g_Config.rendering.enable_lighting },
     { nullptr, nullptr },
 };
 
@@ -38,3 +39,4 @@ REGISTER_CONSOLE_COMMAND("cheats", M_Entrypoint, GS_ID(CONSOLE_HELP_CHEATS))
 REGISTER_CONSOLE_COMMAND(
     "wireframe", M_Entrypoint, GS_ID(CONSOLE_HELP_WIREFRAME))
 REGISTER_CONSOLE_COMMAND("fps", M_Entrypoint, GS_ID(CONSOLE_HELP_FPS))
+REGISTER_CONSOLE_COMMAND("lighting", M_Entrypoint, GS_ID(CONSOLE_HELP_LIGHTING))