+++ docs/tr2/CHANGELOG.md
@@ -5,8 +5,9 @@
 - added the ability to skip end game credits (#3266)
 - added the ability to hide specific game settings (#3242)
 - added the ability to cycle UI tabs with sidestep keys (#3272)
-- added a new easter egg command
 - added dedicated British English translation (#3212)
+- added a new easter egg command
+- added a `/lighting` console command to let the player turn lighting system on/off
 - changed controls dialog to remember the player's preferred input method
 - changed UI to show icons relevant to the chosen input method
 - changed death timer skip to only trigger with Action and Inventory keys